refactor(agent-receipt): move the G5 proposal receipts to capabilityCore
阶段 4 前置 PR ②（方案 docs/plan/2026-09-07-agent-rebuild-stage3-5-deep-plan.md
§2.2 G5 行 / §2.4）。零行为变化：git mv 四个实现 + 三个测试，文件内容除 import
路径外一字未改（26 增 / 26 删，全是 import 行）。

搬家理由（§2.2 G5）：这四件不是 Agent 转录，是**领域收据**——画布提案的撤销程序、
崩溃后「批过的动作到底写没写」的核对依据，且有活的非 Agent 调用者
（mcpStdioServer.ts、rpcServer.ts、mcpDocumentWriteReceipt.ts）。阶段 4 删
projectAgentHost/ 时它们必须活下来，所以 owner 先搬到 capabilityCore/。

落盘格式一字未动：<project>/.nomi/project-agent-proposal-receipt.json、
schemaVersion 2、键顺序、缩进、结尾换行、两个摘要域串全部由上一条 commit 的
金样本测试钉住——同一个金样本在搬家前后都绿。
