+++ electron/capabilityCore/projectAgentProposalReceiptFormat.test.ts
