+++ electron/capabilityCore/projectAgentProposalReceiptStore.test.ts
