+++ electron/capabilityCore/projectAgentReceiptResolver.test.ts
